gta9wifi: Initial recovery tree configuration.
Import initial configuration built with reference to the A15 5G recovery tree. With the initial organization of the recovery tree itself and the touchscreen fix on their device.

We can boot to TWRP, that can register some touch events but does not fully work.

Big props to notes on approach to patching non-functional MTK touch drivers as well.
https://github.com/ADeadTrousers/twrp_device_Unihertz_Atom_LXL/blob/master/docs/HOW-TO-PATCH.md

Co-authored-by: devhunter1 <anubhabkundu2@gmail.com>
Co-authored-by: lopestorm <guillopes2002@gmail.com>
Co-authored-by: Matthias Leitl <a.dead.trousers@gmail.com>
Signed-off-by: clexan <clexan.priv@protonmail.com>